pyqt5/pyqt5_yolov5_detect/requirements.txt

matplotlib==3.5.3
Nuitka==1.3.3
numpy==1.21.5
onnx==1.12.0
onnx-simplifier==0.4.8
onnxruntime==1.12.1
# opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
orderedset==2.0.3
# pandas==1.3.5

PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.6
# PyQtWebEngine-Qt5==5.15.2
PySide6==6.4.1
# PySide6-Addons==6.4.1
# PySide6-Essentials==6.4.1
# pytesseract==0.3.10
# pytest==7.1.3
# sarpy==1.3.10
# scikit-learn==0.22.2
# scipy==1.4.1
# torch==1.10.1
# torchvision==0.11.2
# tqdm==4.64.0
Pillow==9.2.0
